Make VR sample page names match dashboard
Replaces characters that are unsupported by the performance dashboard
with underscores in benchmarks that use WebVR and WebXR sample pages.
This is so that the name reported on the dashboard can actually be
used as a story filter when bisecting.

Bug: 881017
Change-Id: Ifae260815a1382ce5dcab9e393ad8b0344ab0f86
Reviewed-on: https://chromium-review.googlesource.com/c/1252848
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#596025}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3c9032943a2e91d41ac26b976cb620622e04b4cf