Create the HTTPS Server Previews Infobar
This infobar notifies existing DataSaver users that this new feature
will impact their secure page browsing as well. This is done on the
first eligible pageload after the feature is released and enabled. A
simple boolean pref keeps track of this state persistently.

The user does not need to interact with the infobar. Once it has been
shown one time it should not be shown again and this preview can be
triggered. Prior to being shown, the preview should not be triggered
on otherwise eligible page loads.

Users who do not have DataSaver enabled should not see the infobar.

Bug: 889124
Change-Id: I6eb52f516555980a0af90505637201669451c03f
Reviewed-on: https://chromium-review.googlesource.com/c/1244903
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#596131}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 21bafcb21070ff7ba2c4f5ae32d93780c159ecc2