Make histograms/pretty_print_test.py run on the bots.
This test was added via:
https://chromium-review.googlesource.com/1244553

This is a follow-up CL to add it to metrics_python_tests so
that it will run on the bots.

Renames Python files to avoid name collisions between
different scripts, so that they can run together as part of
metrics_python_tests. Also fixes up the rappor model test
and adds it to the suite.

Bug: 714123
Change-Id: I466584493f7090decdb8b21217347a905b3e6391
Reviewed-on: https://chromium-review.googlesource.com/c/1255596
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#596921}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5c5ce638e36966b872f26e2df86116388e06fd51