Report readonly password fields with UKM
When Chrome parses FormData of password forms into PasswordForm
structures, it omits, under certain conditions, password fields marked
as readonly. To understand the potential impact of removing this
special handling, this CL adds a UKM metric to report when this
occurs. This will provide both the frequency and some sample sites to
understand the usage of readonly password fields.

Privacy review of this metric:
The whole UKM event "PasswordForm" has been reviewed in crbug.com/728707
and go/gkmnc, where the privacy TL agreed that adding new metrics for
this event is OK under certain conditions (see the linked doc). The
new twin metrics for the readonly attribute satisfy such conditions, in
the opinion of the CL author, and hence are covered by that review. The
new metrics have been added to the requested spreadsheet listing all
passwords-related metrics.

Choice of metrics:
The core of the metrics is the ReadonlyPasswordFields enum, which
describes different situations which can happen wrt. readonly fields
during parsing. Those are in the UKM combined with the bit indicating
whether the parsing overall was successful, because this is also
relevant to the issue.
Further, the metrics can be collected at two stages: when a form is
parsed to be filled, and when it is parsed to be saved. Here there
was a choice to either record this as an additional bit in one metric,
or create two "twin" metrics. The advantage of the former is the
ability to study how these two events correlate. The advantage of the
latter is easier reading of the particular metrics and a lower range
of values per metric. Because the correlation is unlikely to be of
significant use, the latter approach was chosen.

Bug: 883633
Change-Id: Ib63e52908ae3a257546646c1b255a6cd8fd48b6b
Reviewed-on: https://chromium-review.googlesource.com/c/1261984
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#597561}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 845ecd2b79f187f227f79ced20885f3f1340334e