Deprecate "Smart Lock" in the native UI on desktop.
The CL changes the layout of the "Save username?" bubble. It is used for
federated credentials.
The "Save password/username?" bubble has always a footer for the signed-in users.
The footer is now shown in the account chooser for the signed-in users.

Bug: 890336
Change-Id: I7133cb15c6377e82e971e793886f8520b5bfd4eb
Reviewed-on: https://chromium-review.googlesource.com/c/1265259
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Commit-Queue: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#597565}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a98a17c068a33fac936c1eecde53fd82ba2781f