[Extensions] Remove streamsPrivate API
Remove the streamsPrivate API. It used to be needed by both the PDF
Reader extension and QuickOffice, but both have since migrated to the
MimeTypesHandler API.

Remove the API schema and function implementations, as well as the
feature declarations. Deprecate the old enum values used for the
API calls and permissions.

We need to leave in a portion of the StreamsPrivateAPI implementation
(specifically, StreamsPrivateAPI::SendExecuteMimeTypeHandlerEvent())
because it is also used in the MimeTypesHandler API. I've added a TODO
to move it to a new location, since streamsPrivate no longer makes
sense for it.

Bug: 890401

Change-Id: Idfb0d2504efc5f0a3ad269ae59d836d8e1d598ca
Reviewed-on: https://chromium-review.googlesource.com/c/1252322
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Zachary Kuznia <zork@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#597972}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4750dc8238e821b924eb17470d0714a553dd007b