Change HTTPS Preview handling of 307 and 403
The server has been reconfigured to send 307s for a bypass signal and
403s were also added to signal a bad chrome-proxy header. This CL
updates the client to count the proper metrics for these signals.

Note that there are two types of 307s, a bypass and a normal redirect.
If the Location header points to the original URL, it is a bypass. Else,
not.

Also makes the browser tests a bit easier to read given the change.

Bug: 864656
Change-Id: I20a782fb80feb1135a26d7f47e45e811954bca83
Reviewed-on: https://chromium-review.googlesource.com/c/1270055
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#598148}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d6f501e4e0c15037cafbb035c95f8e1cab8e9153