Revert "Set symbol_level=0 on Windows Deterministic."
This reverts commit 4e9c4b371d71f4b6fe73fce2272cf4e31d11e250.

Reason for revert: This might no longer be needed; revert to check.

(I thought the bot built in two different build dirs, and for that
to work we need to land https://chromium-review.googlesource.com/c/chromium/src/+/1273479
first, but it builds in one dir, moves that elsewhere, and builds
in that dir again. We should probably make it build in two different
dirs once that other CL is in, to make sure that that keeps working.)

Original change's description:
> Set symbol_level=0 on Windows Deterministic.
> 
> If /debug is passed to lld, the pdb aging in PDB70DebugInfo currently
> means that repeat builds get different PDB70DebugInfo headers, so for now
> don't pass /debug to the linker.
> 
> Bug: 330260
> Change-Id: I86e6b470ad95f3f6556c8f8f69d6d394a32661ff
> Reviewed-on: https://chromium-review.googlesource.com/1162948
> Reviewed-by: Dirk Pranke <dpranke@chromium.org>
> Commit-Queue: Nico Weber <thakis@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#580765}

TBR=thakis@chromium.org,dpranke@chromium.org

# Not skipping CQ checks because original CL landed > 1 day ago.

Bug: 330260
Change-Id: I8637e6c089464eddc7d9cfc35f6fa28b0cd54987
Reviewed-on: https://chromium-review.googlesource.com/c/1277486
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#599056}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f8ebcecf7a1b5a9d628ed8dd81727da65fff7ecc