Manually load the original page when ReloadType::ORIGINAL_REQUEST_URL
This is a manual workaround so that HTTPS Server Previews can be team
fooded in 71 and will need to be merged back into the M71 branch.

This checks if the user is on a preview page and is reloading via
clicking the "load original" UI and if so, loads the orignal page.

Bug: 864187
Change-Id: I2c8ed4edba98f30f398675c38d7f89d3848a0212
Reviewed-on: https://chromium-review.googlesource.com/c/1280853
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#599699}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0860cf6232cfa069214f15f4fac77aba9248c6a