Add accuracy metrics input event prediction
This CL adds histogram metrics to measure input prediction accuracy,

There are 3 metrics for each of mouse/touch/scroll (9 total). They
are measuring the distance between predicted position at event time
to the event's position.
We separate each of the result in 3 metrics base on the prediction
time delta, because the accuracy is sensitive to how far we are
predicting.

This CL changes the InputEventPrediction and ScrollPredictor class
to choose different predictor by the *PredictorTypeChoice flag, and
do the prediction when the flag passing corresponding predictor type,
so we will gather accuracy data for the predictor via finch.
Whether apply the resampling to the event is still under the resampling
feature flag.

Bug: 836352
Change-Id: Ie66e4fe1c810120415f9ee976ffa4d8faa0003f1
Reviewed-on: https://chromium-review.googlesource.com/c/1237240
Commit-Queue: Ella Ge <eirage@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#600219}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2cf1bb0ffbf773b884b31df0b6ab0237acd45a05