[Background Fetch] Simplify fetch storage workflow.
In addition, MatchAll should return all records, including unprocessed
ones. To support this, the following changes were made:
- Creating a registration also stores all the requests with an empty
response in the cache.
- When an individual request is processed, the failure reason (if any)
is stored in the metadata.
- The logic of GetSettledFetchesTask was moved to
MarkRegistrationForDeletionTask. It checks the metadata rather than the
cache itself to find a failure reason (if any).
- Match/MatchAll logic was moved to a new database task
(MatchRequestsTask). A new API call was added to the cache storage to
allow querying request/response pairs. If a response is found to be
empty it will be exposed as a nullptr.

Change-Id: I631a3ef3da95117aed759a675fe591da5201eeca
Reviewed-on: https://chromium-review.googlesource.com/c/1280851
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Commit-Queue: Rayan Kanso <rayankans@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#600768}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ac0b91e2c70d357caee4d41bf63967e9b773f4e9