HTTPS Previews: Blacklist hosts when directed by server
If the previews server responds with a bypass (307) and
`chrome-proxy: host-blacklisted` then the original host
should be blacklisted from previews for 30 days.

This information is stored in prefs.

Also adds a histogram to track whether or not a bypass is blacklisted.

Bug: 864649
Change-Id: I42012e7ed3d2eb01d0233636ad8bac0202351946
Reviewed-on: https://chromium-review.googlesource.com/c/1297086
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#602899}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a2b08947596e1510f3a6fa2f65c2491eeeb4e9ec